Add new JAR file
JAR file now runs Duke that has a JavaFX-created GUI instead
of the previous command line interface.
