+++ src/main/java/dukeprotemp/DialogBox.java
@@ -1,0 +1,52 @@
+package dukeprotemp;
+
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.geometry.Insets;
+import javafx.geometry.Pos;
+import javafx.scene.Node;
+import javafx.scene.control.Label;
+import javafx.scene.image.ImageView;
+import javafx.scene.layout.Background;
+import javafx.scene.layout.HBox;
+import javafx.scene.shape.Circle;
+
+public class DialogBox extends HBox {
+
+    private Label text;
+    private ImageView displayPicture;
+
+    public DialogBox(Label l, ImageView iv, Background background) {
+        l.setPadding(new Insets(0, 10.0, 0, 10.0));
+        text = l;
+        displayPicture = iv;
+        text.setWrapText(true);
+        displayPicture.setFitWidth(100.0);
+        displayPicture.setFitHeight(100.0);
+        //set to circle
+        Circle clip = new Circle(50, 50, 50);
+        displayPicture.setClip(clip);
+
+        this.setBackground(background);
+        this.setAlignment(Pos.TOP_RIGHT);
+        this.getChildren().addAll(text, displayPicture);
+        this.setPadding(new Insets(10.0, 15.0, 10.0, 15.0));
+    }
+
+    private void flip() {
+        this.setAlignment(Pos.TOP_LEFT);
+        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
+        FXCollections.reverse(tmp);
+        this.getChildren().setAll(tmp);
+    }
+
+    public static DialogBox getUserDialog(Label l, ImageView iv, Background background) {
+        return new DialogBox(l, iv, background);
+    }
+
+    public static DialogBox getDukeDialog(Label l, ImageView iv, Background background) {
+        var db = new DialogBox(l, iv, background);
+        db.flip();
+        return db;
+    }
+}
