+++ src/main/java/dukeprotemp/DukePro.java
@@ -1,0 +1,157 @@
+package dukeprotemp;
+
+import duke.*;
+import javafx.application.Application;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.scene.layout.*;
+import javafx.scene.paint.Color;
+import javafx.stage.Stage;
+
+
+public class DukePro extends Application {
+
+
+    private ScrollPane scrollPane;
+    private VBox dialogContainer;
+    private TextField userInput;
+    private Button sendButton;
+    private AnchorPane mainLayout;
+    private Scene scene;
+    private Image user = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
+    private Image duke = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
+    private Background userBackground;
+    private Background botBackground;
+
+    //Duke stuff
+    private Storage storage;
+    private TaskList tasks;
+    private Ui ui;
+
+    private Parser parser;
+
+    private Label getDialogLabel(String text) {
+        Label textToAdd = new Label(text);
+        textToAdd.setWrapText(true);
+        return textToAdd;
+    }
+    private void handleUserInput() {
+        try {
+            String response = parser.read(userInput.getText()).execute();
+            Label userText = new Label(youSaid(userInput.getText()));
+            Label dukeText = new Label(getResponse(response));
+            dialogContainer.getChildren().addAll(
+                    DialogBox.getUserDialog(userText, new ImageView(user), userBackground),
+                    DialogBox.getDukeDialog(dukeText, new ImageView(duke), botBackground)
+            );
+        } catch (DukeException e) {
+            Label dukeText = new Label(getResponse(e.getMessage()));
+            dialogContainer.getChildren().addAll(
+                    DialogBox.getDukeDialog(dukeText, new ImageView(duke), botBackground)
+            );
+        }
+        userInput.clear();
+    }
+
+    /**
+     * You should have your own function to generate a response to user input.
+     * Replace this stub with your completed method.
+     */
+    private String getResponse(String input) {
+        return "Duke says: " + input;
+    }
+
+    private String youSaid(String input) {
+        return "You: " + input;
+    }
+
+    public void startupDuke() {
+        ui = new Ui();
+        storage = new Storage("src/dukesave/Duke.txt");
+        try {
+            tasks = new TaskList(storage.load());
+        } catch (DukeException e) {
+            ui.showLoadingError(e);
+            tasks = new TaskList();
+        }
+        parser = new Parser(ui, tasks, storage);
+    }
+
+    public void greet() {
+        Label dukeText = new Label(getResponse(ui.greet()));
+        dialogContainer.getChildren().addAll(
+                DialogBox.getDukeDialog(dukeText, new ImageView(duke), botBackground)
+        );
+    }
+
+    @Override
+    public void start(Stage stage) {
+        scrollPane = new ScrollPane();
+        dialogContainer = new VBox();
+        //set padding
+        dialogContainer.setSpacing(5.0);
+        scrollPane.setContent(dialogContainer);
+        userInput = new TextField();
+        sendButton = new Button("Send");
+        mainLayout = new AnchorPane();
+        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
+        userBackground = new Background(new BackgroundFill(Color.web("#b6f542"), new CornerRadii(30.0), null));
+        botBackground = new Background(new BackgroundFill(Color.web("#f5cb42"), new CornerRadii(30.0), null));
+
+        scene = new Scene(mainLayout);
+        stage.setScene(scene);
+        stage.show();
+
+        stage.setTitle("Duke");
+        stage.setResizable(false);
+        stage.setMinHeight(600.0);
+        stage.setMinWidth(400.0);
+
+        mainLayout.setPrefSize(400.0, 600.0);
+        scrollPane.setPrefSize(385, 535);
+        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
+        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
+
+        scrollPane.setVvalue(1.0);
+        scrollPane.setFitToWidth(true);
+
+        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
+
+        userInput.setPrefWidth(325.0);
+        sendButton.setPrefWidth(55.0);
+
+        AnchorPane.setTopAnchor(scrollPane, 1.0);
+        AnchorPane.setBottomAnchor(sendButton, 1.0);
+        AnchorPane.setRightAnchor(sendButton, 1.0);
+
+        AnchorPane.setLeftAnchor(userInput , 1.0);
+        AnchorPane.setBottomAnchor(userInput, 1.0);
+
+        sendButton.setOnMouseClicked((event) -> {
+            dialogContainer.getChildren().add(getDialogLabel(userInput.getText()));
+            userInput.clear();
+        });
+
+        userInput.setOnAction((event) -> {
+            dialogContainer.getChildren().add(getDialogLabel(userInput.getText()));
+            userInput.clear();
+        });
+
+        sendButton.setOnMouseClicked((event) -> {
+            handleUserInput();
+        });
+
+        userInput.setOnAction((event) -> {
+            handleUserInput();
+        });
+        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
+        startupDuke();
+        greet();
+    }
+}
+
