+++ src/main/java/dukeprotemp/Launcher.java
@@ -1,4 +1,4 @@
-package dukepro;
+package dukeprotemp;
 
 import javafx.application.Application;
 
@@ -7,6 +7,6 @@ import javafx.application.Application;
  */
 public class Launcher {
     public static void main(String[] args) {
-        Application.launch(Main.class, args);
+        Application.launch(DukePro.class, args);
     }
 }
